Move register transfer helpers + tests into module
